+++ apollo-core/src/main/java/com/ctrip/framework/apollo/core/ConfigConsts.java
@@ -19,7 +19,8 @@ package com.ctrip.framework.apollo.core;
 public interface ConfigConsts {
   String NAMESPACE_APPLICATION = "application";
   String CLUSTER_NAME_DEFAULT = "default";
-  String CLUSTER_NAMESPACE_SEPARATOR = "-";
+  String CLUSTER_NAMESPACE_SEPARATOR = "+";
+  String CONFIGMAP_KEY_SEPARATOR = "-";
   String APOLLO_CONFIG_CACHE = "apollo-configcache-";
   String APOLLO_CLUSTER_KEY = "apollo.cluster";
   String APOLLO_META_KEY = "apollo.meta";