+++ CHANGES.md
@@ -6,10 +6,11 @@ Apollo Java 2.4.0
 
 ------------------
 
-* [Feature Support Kubernetes ConfigMap cache for Apollo java, golang client](https://github.com/apolloconfig/apollo-java/pull/79)
 * [Fix the Cannot enhance @Configuration bean definition issue](https://github.com/apolloconfig/apollo-java/pull/82)
 * [Feature openapi query namespace support not fill item](https://github.com/apolloconfig/apollo-java/pull/83)
 * [Add more observability in apollo config client](https://github.com/apolloconfig/apollo-java/pull/74)
+* [Feature Support Kubernetes ConfigMap cache for Apollo java, golang client](https://github.com/apolloconfig/apollo-java/pull/79)
+
 
 ------------------
 All issues and pull requests are [here](https://github.com/apolloconfig/apollo-java/milestone/4?closed=1)